+++ percona/perconacollector/custom_query.go
@@ -20,18 +20,18 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"log/slog"
 	"math"
+	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
+	"github.com/alecthomas/kingpin/v2"
+	cl "github.com/percona/mysqld_exporter/collector"
 	"github.com/prometheus/client_golang/prometheus"
-	"gopkg.in/alecthomas/kingpin.v2"
 	"gopkg.in/yaml.v2"
 )
 
@@ -121,7 +121,7 @@ func (scq ScrapeCustomQuery) Version() float64 {
 }
 
 // Scrape collects data.
-func (scq ScrapeCustomQuery) Scrape(ctx context.Context, db *sql.DB, ch chan<- prometheus.Metric, logger log.Logger) error {
+func (scq ScrapeCustomQuery) Scrape(ctx context.Context, instance *cl.Instance, ch chan<- prometheus.Metric, logger *slog.Logger) error {
 	cq := CustomQuery{
 		customMetricMap: make(map[string]MetricMapNamespace),
 		customQueryMap:  make(map[string]string),
@@ -133,7 +133,7 @@ func (scq ScrapeCustomQuery) Scrape(ctx context.Context, db *sql.DB, ch chan<- p
 		HR: collectCustomQueryHrDirectory,
 	}
 
-	fi, err := ioutil.ReadDir(*dirs[scq.Resolution])
+	fi, err := os.ReadDir(*dirs[scq.Resolution])
 	if err != nil {
 		return fmt.Errorf("failed read dir %q for custom query. reason: %s", *dirs[scq.Resolution], err)
 	}
@@ -145,7 +145,7 @@ func (scq ScrapeCustomQuery) Scrape(ctx context.Context, db *sql.DB, ch chan<- p
 
 		if filepath.Ext(v.Name()) == ".yml" || filepath.Ext(v.Name()) == ".yaml" {
 			path := filepath.Join(*dirs[scq.Resolution], v.Name())
-			userQueriesData, err := ioutil.ReadFile(path)
+			userQueriesData, err := os.ReadFile(path)
 			if err != nil {
 				return fmt.Errorf("failed to open custom queries:%s", err.Error())
 			}
@@ -161,6 +161,7 @@ func (scq ScrapeCustomQuery) Scrape(ctx context.Context, db *sql.DB, ch chan<- p
 
 	cq.mappingMtx.RLock()
 	defer cq.mappingMtx.RUnlock()
+	db := instance.GetDB()
 	errMap := queryNamespaceMappings(ctx, ch, db, cq.customMetricMap, cq.customQueryMap, logger)
 	if len(errMap) > 0 {
 		errs := make([]string, 0, len(errMap))
@@ -197,7 +198,7 @@ func (cm *ColumnMapping) UnmarshalYAML(unmarshal func(interface{}) error) error 
 // addQueries metricMap and customQueryMap to contain the new queries.
 // Added queries do not respect version requirements, because it is assumed that
 // the user knows what they are doing with their version of mysql.
-func addQueries(content []byte, exporterMap map[string]MetricMapNamespace, customQueryMap map[string]string, logger log.Logger) error {
+func addQueries(content []byte, exporterMap map[string]MetricMapNamespace, customQueryMap map[string]string, logger *slog.Logger) error {
 	var extra map[string]interface{}
 	err := yaml.Unmarshal(content, &extra)
 	if err != nil {
@@ -206,7 +207,7 @@ func addQueries(content []byte, exporterMap map[string]MetricMapNamespace, custo
 	// Stores the loaded map representation.
 	metricMaps := make(map[string]map[string]ColumnMapping)
 	for metric, specs := range extra {
-		level.Debug(logger).Log("msg", "New user metric namespace from YAML", "metric", metric)
+		logger.Debug("msg", "New user metric namespace from YAML", metric)
 		specMap, ok := specs.(map[interface{}]interface{})
 		if !ok {
 			return fmt.Errorf("incorrect yaml format for %+v", specs)
@@ -260,7 +261,7 @@ func addQueries(content []byte, exporterMap map[string]MetricMapNamespace, custo
 }
 
 // Turn the MetricMap column mapping into a prometheus descriptor mapping.
-func makeDescMap(metricMaps map[string]map[string]ColumnMapping, exporterMap map[string]MetricMapNamespace, logger log.Logger) {
+func makeDescMap(metricMaps map[string]map[string]ColumnMapping, exporterMap map[string]MetricMapNamespace, logger *slog.Logger) {
 	metricMap := make(map[string]MetricMapNamespace)
 	for namespace, mappings := range metricMaps {
 		thisMap := make(map[string]MetricMap)
@@ -331,7 +332,7 @@ func makeDescMap(metricMaps map[string]map[string]ColumnMapping, exporterMap map
 						case string:
 							durationString = t
 						default:
-							level.Error(logger).Log("msg", "DURATION conversion metric was not a string")
+							logger.Error("msg", "DURATION conversion metric was not a string", durationString)
 							return math.NaN(), false
 						}
 
@@ -341,7 +342,7 @@ func makeDescMap(metricMaps map[string]map[string]ColumnMapping, exporterMap map
 
 						d, err := time.ParseDuration(durationString)
 						if err != nil {
-							level.Error(logger).Log("msg", "Failed converting result to metric", "columnName", columnName, "in", in, "error", err)
+							logger.Error("msg", "Failed converting result to metric", "columnName", columnName, "in", in, "error", err)
 							return math.NaN(), false
 						}
 						return float64(d / time.Millisecond), true
@@ -377,7 +378,7 @@ func stringToColumnUsage(s string) (ColumnUsage, error) {
 
 // Convert "database/sql value" types to float64s for Prometheus consumption.
 // Null types are mapped to NaN. string and []byte types are mapped as NaN and !ok.
-func dbToFloat64(t interface{}, logger log.Logger) (float64, bool) {
+func dbToFloat64(t interface{}, logger *slog.Logger) (float64, bool) {
 	switch v := t.(type) {
 	case int64:
 		return float64(v), true
@@ -390,14 +391,14 @@ func dbToFloat64(t interface{}, logger log.Logger) (float64, bool) {
 		strV := string(v)
 		result, err := strconv.ParseFloat(strV, 64)
 		if err != nil {
-			level.Warn(logger).Log("msg", "Could not parse []byte", "error", err)
+			logger.Warn("Could not parse []byte", "error", err)
 			return math.NaN(), false
 		}
 		return result, true
 	case string:
 		result, err := strconv.ParseFloat(v, 64)
 		if err != nil {
-			level.Warn(logger).Log("msg", "Could not parse string", "error", err)
+			logger.Warn("Could not parse string", "error", err)
 			return math.NaN(), false
 		}
 		return result, true
@@ -432,7 +433,7 @@ func dbToString(t interface{}) (string, bool) {
 // the scrape fails, and a slice of errors if they were non-fatal.
 func queryNamespaceMapping(ctx context.Context, ch chan<- prometheus.Metric,
 	db *sql.DB, namespace string, mapping MetricMapNamespace,
-	customQueries map[string]string, logger log.Logger) ([]error, error) {
+	customQueries map[string]string, logger *slog.Logger) ([]error, error) {
 	// Check for a query override for this namespace.
 	query, found := customQueries[namespace]
 
@@ -481,7 +482,7 @@ func queryNamespaceMapping(ctx context.Context, ch chan<- prometheus.Metric,
 		for idx, columnName := range mapping.labels {
 			labels[idx], ok = dbToString(columnData[columnIdx[columnName]])
 			if !ok {
-				level.Info(logger).Log("msg", "converted NULL to an empty string")
+				logger.Error("converted NULL to an empty string")
 			}
 		}
 
@@ -527,7 +528,7 @@ func queryNamespaceMapping(ctx context.Context, ch chan<- prometheus.Metric,
 // Iterate through all the namespace mappings in the exporter and run their queries.
 func queryNamespaceMappings(ctx context.Context, ch chan<- prometheus.Metric,
 	db *sql.DB, metricMap map[string]MetricMapNamespace, customQueries map[string]string,
-	logger log.Logger) map[string]error {
+	logger *slog.Logger) map[string]error {
 	// Return a map of namespace -> errors.
 	namespaceErrors := make(map[string]error)
 	for namespace, mapping := range metricMap {
@@ -539,7 +540,7 @@ func queryNamespaceMappings(ctx context.Context, ch chan<- prometheus.Metric,
 		// Non-serious errors - likely version or parsing problems.
 		if len(nonFatalErrors) > 0 {
 			for _, err := range nonFatalErrors {
-				level.Info(logger).Log(err.Error())
+				logger.Info(err.Error())
 			}
 		}
 	}